requirements.txt
playwright==1.50.0
playwright-stealth==2.0.2
google-genai>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
